+++ EasyLinalg/docs/Exemplos/arrow2d_aggregated.png
