+++ EasyLinalg/docs/Exemplos/arrow3d_aggregated.png
