+++ EasyLinalg/docs/Exemplos/signal_aggregated.png
