+++ EasyLinalg/docs/Exemplos/signal_separated.png
