+++ EasyLinalg/test/image/bola.jpg
