+++ EasyLinalg/test/image/quadrado.jpg
