+++ EasyLinalg/test/runtests.jl
@@ -1,0 +1,10 @@
+using Test
+using EasyLinalg
+
+include(joinpath("signal", "signalTeste.jl"))
+include(joinpath("point2D", "point2DTeste.jl"))
+include(joinpath("point3D", "point3DTeste.jl"))
+include(joinpath("arrow2D", "arrow2DTeste.jl"))
+include(joinpath("arrow3D", "arrow3DTeste.jl"))
+# include(joinpath("image", "imageTeste.jl"))
+
